requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-dotenv
python-multipart # For form data/file uploads
eval_type_backport  # For Python <3.10 compatibility

# LLM Integration
openai>=1.5.0  # For OpenRouter API
instructor>=0.5.0  # For structured outputs

# Document parsing
PyMuPDF
python-docx
aiofiles

# Testing
pytest
httpx
pytest-asyncio
pytest-env

datetime
sse-starlette
